+++ color.h
@@ -1,0 +1,26 @@
+#ifndef COLOR_H
+#define COLOR_H
+
+typedef enum _color_t {
+    BLACK, //0
+    BLUE, // 1
+    GREEN,
+    CYAN,
+    RED,
+    PURPLE,
+    YELLOW,
+    WHITE,
+    GRAY,
+    LIGHT_BLUE,
+    LIGHT_GREEN,
+    LIGHT_CYAN,
+    LIGHT_RED,
+    LIGHT_PURBPLE,
+    LIGHT_YELLOW,
+    LIGHT_WHITE,
+} color_t;
+
+
+char* get_color_string(color_t color);
+
+#endif
