+++ main.c
@@ -2,9 +2,10 @@
 #include <stdint.h>
 #include <stdbool.h>
 #include <windows.h>
-//test
+
 #include "blocks.h"
 #include "key_driver.h"
+#include "color.h"
 
 #define _CRT_SECURE_NO_WARNINGS
 
@@ -14,55 +15,6 @@
 #define RESIZE_COLS (DEFAULT_COLS-30) //=90
 #define RESIZE_ROWS 30
 
-//추가적으로 밝은 색상은 해당 값에 8을 더한 값으로 표현됩니다(예: 밝은 파랑은 9).
-typedef enum _color_t {
-    BLACK, //0
-    BLUE, // 1
-    GREEN,
-    CYAN,
-    RED,
-    PURPLE,
-    YELLOW,
-    WHITE,
-    GRAY,
-    LIGHT_BLUE,
-    LIGHT_GREEN,
-    LIGHT_CYAN,
-    LIGHT_RED,
-    LIGHT_PURBPLE,
-    LIGHT_YELLOW,
-    LIGHT_WHITE,
-} color_t;
-
-
-//추가적으로 밝은 색상은 해당 값에 8을 더한 값으로 표현됩니다(예: 밝은 파랑은 9).
-char* color_strings[16] = {
-    "black", // 검정
-    "blue", // 파랑
-    "green", // 초록
-    "cyan", // 청록
-    "red", // 빨강
-    "purple", // 자주
-    "yellow", // 노랑
-    "white", // 흰색
-    "gray", // 회색
-    "light blue", // 밝은 파랑
-    "light green", // 밝은 회색
-    "light cyan", // 밝은 청록
-    "light red", // 밝은 빨강
-    "light purple", // 밝은 자주
-    "light yellow", // 밝은 노랑
-    "light white" // 밝은 흰색
-};
-
-char* get_color_string(color_t color) {
-    if (color >= BLACK && color <= LIGHT_WHITE) {
-        return color_strings[color]; // 유효한 색상 값이면 해당 문자열 반환
-    }
-    else {
-        return "unknown"; // 유효하지 않은 값이면 기본값 반환
-    }
-}
 
 typedef struct _cursor_info_t {
     uint32_t x;
@@ -325,7 +277,7 @@ void console_clear(const windows_console_t* console) {
     FillConsoleOutputAttribute(hConsole, csbi.wAttributes, console_size, top_left, &chars_written); // 콘솔 화면의 모든 속성을 기본값으로 채우기
     SetConsoleCursorPosition(hConsole, top_left); // 커서를 화면의 좌상단으로 이동
 }
-//color
+
 void setcolor(unsigned short text, unsigned short back) {
     SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), text | (back << 4));
 }
